feat: feat/update-helm-functions
