+++ helm-old/.gitignore
@@ -2,4 +2,3 @@
 /internal/dagger
 /internal/querybuilder
 /internal/telemetry
-/.env
