+++ helm-old/dagger.json
@@ -1,0 +1,16 @@
+{
+  "name": "helm",
+  "sdk": "go",
+  "dependencies": [
+    {
+      "name": "helm-disaster37",
+      "source": "github.com/disaster37/dagger-library-go/helm@2d35aed06292796c4d2351b6336caa1f37247f7b"
+    },
+    {
+      "name": "helm-oci",
+      "source": "github.com/purpleclay/daggerverse/helm-oci@d7438770bfab8844a89c2923b9e2942e78de5239"
+    }
+  ],
+  "source": ".",
+  "engineVersion": "v0.13.3"
+}
